Unify vkb::Gui and vkb::HPPGui into vkb::Gui<bindingType>
