+++ framework/hpp_api_vulkan_sample.h
@@ -19,7 +19,6 @@
 
 #include <camera.h>
 #include <common/hpp_error.h>
-#include <hpp_gui.h>
 #include <scene_graph/components/hpp_image.h>
 #include <scene_graph/components/hpp_sub_mesh.h>
